requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==3.0.0
fsspec==2023.12.2
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
requests==2.31.0
sympy==1.12
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==2.1.2
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torchvision==0.16.2
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
